+++ src/backend/dotnet/Services/Portal/Freaks.Portal.SharedContracts/Dto/RaidSummary/RaidParticipantDto.cs
@@ -9,5 +9,5 @@ namespace Freaks.Portal.SharedContracts.Dto.RaidSummary;
 /// <param name="RaidId">Идентификатор рейда, в котором участвует пользователь.</param>
 /// <param name="Participant">Информация о пользователе, участвующем в рейде.</param>
 public record RaidParticipantDto(
-    int RaidId,
+    long RaidId,
     UserDto Participant);
